[z3tracer] further restrict APIs and create model.rs
